component renamed, rout updated
