+++ FrontEnd/src/pages/Authorization/LoginPage.module.css
@@ -2,7 +2,7 @@
   display: flex;
   justify-content: center;
   width: 100vw;
-  min-height: var(--min-height-block-main);
+  flex-grow: 1;
   flex-shrink: 0;
   background: var(--light-seashell-background);
   padding: 80px 0;
